Prevent infinite loop when collapsing a node
Use case to reproduce the old bug follows. Have a node X which is under
node Y. When you want to collapse path (X,) instead of (Y,X), GTG burns
CPU a lot. It finds node X but uses original path what means it don't
find it and it has to be done again.