debian: Switch dh build system to pybuild
Pybuild[1] is the recommended[2] dh build system for Python 3 modules,
as python_distutils only supports Python 2.

As a bonus, since pybuild builds within a dedicated `.pybuild` directory
which it removes on cleanup, override_dh_auto_clean is no longer needed
and can be removed.

Also note that Build-Depends-Indep was folded back within Build-Depends,
as it is mostly useless for arch:all packages.

(Testing is disabled for the moment; the next commit will fix this.)

 [1] https://wiki.debian.org/Python/Pybuild
 [2] https://wiki.debian.org/Python/LibraryStyleGuide#Overview