Expliciting LGPLv3-or-later in the README file
I researched and found that to license the source code under LGPLv3-or-later
you just need to specify it in the source code header, but the license is the
same so I don't need to make any changes to the LICENSE file and the source
files already have the explicit "or any later version" in the header.

I'm just modifying the README file because Github doesn't have a way to tell
the person visiting the repository that the program is LGPLv3-or-later, it
just says LGPLv3, so I think the most effective way to specify it is to put
it in the README file.